+++ Dev/SourcesEngine/Gugu/Element/2D/ElementText.h
@@ -52,7 +52,7 @@ public:
 
     void SetFont        (const std::string& _strFontPath);
     void SetFont        (Font* _pFont);
-    void SetFontSize    (size_t _uiSize);
+    void SetFontSize    (unsigned int _uiSize);
     void SetFontColor   (const sf::Color& _oColor);
     //TODO: Add accessors for all sf::Text methods (LineSpacing, LetterSpacing, Style, etc)
